021-conical-continued-again/magnum_opus/requirements.txt
appdirs
flask
flask-restful
flask-marshmallow
flask-sqlalchemy
marshmallow-sqlalchemy
sqlalchemy-utils
